+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 02/06/2022
+### add
+- TransactionDetail component
+- Collapse component
+### update
+- Fix color props Button
+### delete
+- Header component deleted
+
 ## [0.5.1] - 11/05/2022
 ### delete
 - Automplete component